backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docker==7.1.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
numpy==2.3.2
openai==1.98.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pip-tools==7.5.0
playwright==1.54.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
